refactor(sanity): make single bundle menu + clean up
